+++ src/bricks/Empty.hpp
@@ -15,7 +15,7 @@ Empty<Shape>::~Empty() {}
 
 template <typename Shape>
 void Empty<Shape>::renderCell(std::shared_ptr<SDL_Renderer>& renderer, int x,
-                              int y, int cellWidth, int cellHeight) {
+                              int y, float cellWidth, float cellHeight) {
   this->shape_.draw(renderer, x, y, cellWidth, cellHeight,
                     ColorUtils::convertColor(getColor()));
 }
